Backend_Compile_Club/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
environ==1.0
PyJWT==2.10.1
sqlparse==0.5.3
whitenoise==6.9.0
